requirements.txt
requests==2.29.0
python-dotenv==1.0.0
openai==0.27.6